Move event_stream to event_manager api and impl modules
